+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Unreleased
 
 - Remove `PartialOrd` bound for `UserEvent`.
 - Add `Send` bound for `UserEvent` to trait `Poll`, as was already required for adding it to `SyncPort`.
+- Remove unnecessary bounds on Input Event Listeners.
 
 ## 3.0.1